Update project_text.js
I changed the sizes of the graphics and made the bullet shoot out of the turtle's head so its more clear for the user the direction of aim. I started working on the level infustruture, though I have been having a hard time getting it to work after playing with it for a while. I made the canvas bigger and spent some time figuring out how to get the bullets to go farther- has to do with the life of the bullets. Having it go farther by increasing the life had this effect of the bullets wrapping beyond the border, which I think I like, but I'm not sure. I'm also struggling with trying to get everything to stary within the screen because now the cans leave the canvas.